refactor: add missing imports (part 4)
