+++ src/renderer/app/models/tab.ts
@@ -4,6 +4,12 @@ import { ipcRenderer } from 'electron';
 
 import store from '~/renderer/app/store';
 import { ISite } from '~/interfaces';
+import {
+  defaultTabOptions,
+  ADD_TAB_MARGIN_LEFT,
+  TABS_PADDING,
+  TAB_ANIMATION_DURATION,
+} from '../constants/tabs';
 
 let id = 1;
